Added the first fogdrive sources from a private SVN repo.
The first FogDrive, Mira, is under development.

So far, it has a debounced switch which is directly used to activate/decative the vaporizer.
The basic architecture ideas are in place and the preconditions of building it for multiple
architecture should also be there. The build system is set up with SCons. A simple framework
for variant management is implemented.
The main focus now is on battery voltage measurement, a more sophisticated LED control and
a more sophisticated switch control as well as a port to a smaller μC.

The sources have some residual code from some battery voltage measurement experiments. That will be cleaned soon.

Good luck, FogDrive!